app/requirements.txt
flask
keras
numpy
pillow
gunicorn
flask-monitoringdashboard
